Fix #85: empty CURIE must be made illegal
I checked the reference and cmaeder is right, the ISO/IEC 14977:1996
standard we reference in the Normative References chapter defines "-" as
the "except-symbol" with exactly the described meaning. So I went ahead
and changed the CURIE definition using "-".  I originally was going to
use additional nonterminals instead but the solution turned out to be
way less readable than I thought. For a comparison, here is the solution
I came up with:

  CURIE   ::= CwithPath | CnoPath ;
  CwithPath ::= [ Prefix ] , Reference ;
  CnoPath ::= CPrefix | CQuery | CFragment ;
  CPrefix ::= Prefix , EPath , [ Query ] , [ Fragment ] ;
  CQuery  ::= [ Prefix ] , EPath , Query , [ Fragment ] ;
  CFragment ::= [ Prefix ] , EPath , [ Query ] , Fragment ;
  Reference ::= Path , [ Query ] , [ Fragment ] ;
  Path    ::= ipath-absolute | ipath-rootless ; ($<$ \rm as defined in \nisref{IETF/RFC 3987} $>$)
  EPath   ::= ipath-empty ; ($<$ \rm as defined in \nisref{IETF/RFC 3987} $>$)

That is, I only split up Path into an empty and a nonempty version.
For nonempty Paths the CURIE definition can stay the same while for
empty Paths I just specified three alternatives, each leaving one of
Prefix, Query or Fragment non-optional. All unmentioned definitions stay
the same.
Note that one could also use more explicit nonterminal names like
CURIEwithNonEmptyPath or CURIEwithNonOptionalPrefix and so on but I
doubt that would help readability. This grammar is also ambiguous as
CPrefix, CPath and CFragment share a portion of the parse tree. I don't
know whether this is a problem, but a solution avoiding this would be a
lot more complex (I think).

If any reviewer still deems the above more desirable than using a
possibly not widely known feature of ISO-EBNF in two places, feel free
to reopen the issue and I'll switch to the more verbose solution.