Address #86
Is there any problem with simply including '#' into the allowed
characters for separators? Because that is the solution I went for. Note
that '#' is part of "gen-delims" in RFC 3987 and we already include '/'
and '?', which are also "gen-delims", as allowed separator characters.
So an alternative solution would be to have:

  SeparatorChar = ipchar | gen-delims ;

thus getting "ipchar"s together with '/', '?', '#', '[', ']', '@', ':'
(where the last two are also covered by "ipchar") as allowed
separator characters. If that is more desirable, I'll change it
accordingly.

I also added a note to 10.2.1 pointing to where "String" is defined, as
I saw the same has been done for "Prefix".