Use 'gen-delims' to specify separator characters
The '#' should be disallowed in separator strings, because using '#' can
have very severe consequences for IRIs. On the other hand one can allow
a few more special characters, which this patch does by just specifying
valid separator characters to be "ipchar"s or "gen-delims" excluding
'#'.
Should time prove that the '#' is wanted as a separator one can always
amend the standard to allow it. The other way round, i.e. forbidding it
later, would be much more difficult or even impossible due to backwards
compatibility.