Fix issue #161: define "lang:" prefix in examples
Apparently the "lang:" prefix is used in a few (three) examples without
being defined. If I read the report correctly that should be a prefix of
IRIs pointing to machine readable definitions of OMS conforming
languages. As I'm not aware of any existing IRIs for this, I just used a
OMG mock up like the ones used for the "log:" and "trans:" prefixes
which are mock ups, (i.e. they don't exist) too.
Please reopen if this is not the right way to go about this issue.

On a side note, there's also a "ser:" prefix used but not defined in a
few of the examples. I don't know whether that's a problem too, or
whether the "ser:" prefix is special as it is used for
syntax/serializations and points to a "special" Hets builtin (as per
comments).