Try to align `InterpretationDefinition` with abstract syntax
It's highly weird, as `RefinementDefinition` and
`InterpretationDefinition` are two distinct and seemingly unrelated
classes in the abstract syntax, whereas in the concrete syntax
`InterpretationDefinition` may also define a refinement, so the two
distinct abstract concepts somehow map the same concrete concept, just
written differently.
The confusion is magnified by the fact that three keywords may be used
interchangeably in the concrete syntax so two distinct abstract
"nonterminals" map the same set of concrete keywords.
I tried to make these things a little bit clearer in this commit but
don't know whether it succeeded.